% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{animateOutput}
\alias{animateOutput}
\title{Create an animate output (container) element}
\usage{
animateOutput(
  outputId = "animateOutput",
  width = "100\%",
  height = "400px",
  ...
)
}
\arguments{
\item{outputId}{output variable to read the plot/image from.}

\item{width}{Width of the plot area. Must be a valid CSS unit (like "100\%", "400px", "auto").}

\item{height}{Height of the plot area. Must be a valid CSS unit (like "100\%", "400px", "auto").}

\item{...}{Optional CSS styling for the container of the plot.}
}
\description{
Create an animate output (container) element
}
\note{
(Advanced usage) A "stack_limit" parameter can be included in the optional parameters
to control how many directives the device should keep track of.
}
