% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{click_to_play}
\alias{click_to_play}
\title{Click an element to play a frame}
\usage{
click_to_play(selector = "#SVG_1", start = 2)
}
\arguments{
\item{selector}{The ID of the DOM element.}

\item{start}{An integer; the number of frames to execute upon the beginning
of the visualisation. This is useful when one wants to start with some set-up
instead of an empty canvas.}
}
\description{
Click an element to play a frame
}
