% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{compile_data}
\alias{compile_data}
\title{#' Loop through the available frames n times
#' @param times An integer; the number of times to loop.
#' @param wait A number; the number of milliseconds to wait for before the
#' next frame is drawn.
#' @export
loop <- function(times = 1, wait = 20) {
sprintf(
"ctrl = new controller(JS_device);
ctrl.loop(\%s, \%s);",
times, wait
)
}}
\usage{
compile_data(input, output = tempfile(), ...)
}
\description{
#' Loop through the available frames n times
#' @param times An integer; the number of times to loop.
#' @param wait A number; the number of milliseconds to wait for before the
#' next frame is drawn.
#' @export
loop <- function(times = 1, wait = 20) {
sprintf(
"ctrl = new controller(JS_device);
ctrl.loop(\%s, \%s);",
times, wait
)
}
}
