% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{ffmpeg_command}
\alias{ffmpeg_command}
\title{'FFmpeg' command builder}
\usage{
ffmpeg_command(input, output, ...)
}
\arguments{
\item{input}{A character string; the path to the input video file.}

\item{output}{A character string; the path to the output video file.}

\item{...}{Parameters to be passed to the \code{ffmpeg} command-line interface.}
}
\value{
A character string; the 'FFmpeg' command.
}
\description{
'FFmpeg' command builder
}
