% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{insert_animate}
\alias{insert_animate}
\title{Insert an animated plot into an R Markdown document}
\usage{
insert_animate(file, options = click_to_play(), style)
}
\arguments{
\item{file}{The exported plot.}

\item{options}{A character string; the JavaScript to customise the playback options.
Two basic options \code{click_to_play()} and \code{click_to_loop()} have been implemented
for general usage.}

\item{style}{Optional style for the iframe that hosts the visualisation.}
}
\description{
Insert an animated plot into an R Markdown document
}
