% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{load_animate}
\alias{load_animate}
\title{Load the assets of an animate plot into a Shiny app}
\usage{
load_animate(n = 0)
}
\arguments{
\item{n}{The number of commands the plot stack can hold. Use -1 for unlimited
number of commands.}
}
\description{
Load the assets of an animate plot into a Shiny app
}
\note{
This function should be called inside the \code{ui} function of a Shiny app.
Note that this function will work only if the 'www' folder of the Shiny app
contains the relevant JavaScript files. These files can be created using the
function \code{setup_animate}.
}
