% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_id}
\alias{new_id}
\title{A utility function for generating IDs}
\usage{
new_id(x, prefix = "ID", sep = "-")
}
\arguments{
\item{x}{The data that require IDs.}

\item{prefix}{A character string; the prefix to be added to each ID.}

\item{sep}{A character string; the separator to be added between the prefix
and an ID.}
}
\description{
A utility function for generating IDs
}
