% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{plot2}
\alias{plot2}
\title{A web-based graphics device}
\description{
A web-based graphics device

A web-based graphics device
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A handle for the WebSocket connection.}

\item{\code{ready_state}}{The ready state of the connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{plot2$new()}}
\item \href{#method-set_max_stacksize}{\code{plot2$set_max_stacksize()}}
\item \href{#method-off}{\code{plot2$off()}}
\item \href{#method-send}{\code{plot2$send()}}
\item \href{#method-init}{\code{plot2$init()}}
\item \href{#method-bars}{\code{plot2$bars()}}
\item \href{#method-plot}{\code{plot2$plot()}}
\item \href{#method-points}{\code{plot2$points()}}
\item \href{#method-lines}{\code{plot2$lines()}}
\item \href{#method-text}{\code{plot2$text()}}
\item \href{#method-image}{\code{plot2$image()}}
\item \href{#method-set}{\code{plot2$set()}}
\item \href{#method-par}{\code{plot2$par()}}
\item \href{#method-remove}{\code{plot2$remove()}}
\item \href{#method-clear}{\code{plot2$clear()}}
\item \href{#method-import}{\code{plot2$import()}}
\item \href{#method-export}{\code{plot2$export()}}
\item \href{#method-clone}{\code{plot2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$new(width, height, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{An integer; the width in pixels.}

\item{\code{height}}{An integer; the height in pixels.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_stacksize"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_stacksize}{}}}
\subsection{Method \code{set_max_stacksize()}}{
Set the maximum size of the stack
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$set_max_stacksize(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of commands the plot stack can hold. Use
-1 for unlimited number of commands.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-off"></a>}}
\if{latex}{\out{\hypertarget{method-off}{}}}
\subsection{Method \code{off()}}{
Switch off the device; this function closes the WebSocket connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$off()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send commands to device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$send(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to send to the device.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
Initialise a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$init(width = 800, height = 600, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Width of the canvas in pixels.}

\item{\code{height}}{Height of the canvas in pixels.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bars"></a>}}
\if{latex}{\out{\hypertarget{method-bars}{}}}
\subsection{Method \code{bars()}}{
Add points to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$bars(x, y, w, h, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the bars.}

\item{\code{y}}{The y coordinates of the bars.}

\item{\code{w}}{The width of the bars.}

\item{\code{h}}{The height of the bars.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Generic X-Y plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$plot(x, y, type = "p", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the data.}

\item{\code{y}}{The y coordinates of the data.}

\item{\code{type}}{Type of the plot; one of 'p' and 'l'.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points"></a>}}
\if{latex}{\out{\hypertarget{method-points}{}}}
\subsection{Method \code{points()}}{
Add points to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$points(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the points.}

\item{\code{y}}{The y coordinates of the points.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lines"></a>}}
\if{latex}{\out{\hypertarget{method-lines}{}}}
\subsection{Method \code{lines()}}{
Add line segments / paths to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$lines(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the line.}

\item{\code{y}}{The y coordinates of the line.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\if{latex}{\out{\hypertarget{method-text}{}}}
\subsection{Method \code{text()}}{
Add text to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$text(x, y, labels, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the text.}

\item{\code{y}}{The y coordinates of the text.}

\item{\code{labels}}{The text.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image"></a>}}
\if{latex}{\out{\hypertarget{method-image}{}}}
\subsection{Method \code{image()}}{
Add background image to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$image(href, width, height, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{href}}{The link to the image.}

\item{\code{width}}{The width of the image.}

\item{\code{height}}{Th height of the image.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the active device to a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$set(selector)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selector}}{A character vector; ID of the SVG.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\if{latex}{\out{\hypertarget{method-par}{}}}
\subsection{Method \code{par()}}{
Set the graphical parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$par(parameters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{The graphical parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
Remove a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$remove(selector)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selector}}{A character vector; ID of an element.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all elements of the active SVG element
#' @description
#' Perform a group of graphical operations to a plot
#' @param ... Any number of graphical operations.
group = function(...) {
self$send(Message("fn_group", c(...)))
},
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
