% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmd_animate}
\alias{rmd_animate}
\title{In-line rendering of an animated plot in an R Markdown document}
\usage{
rmd_animate(device, ...)
}
\arguments{
\item{device}{The \link{animate} object.}

\item{...}{Optional parameters to pass to \link{insert_animate}.}
}
\description{
In-line rendering of an animated plot in an R Markdown document
}
\note{
This function should only be used in a code chunk of an R Markdown document.
}
