% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{setup_animate}
\alias{setup_animate}
\title{Copy the necessary JavaScript assets to the 'www' folder of a Shiny app}
\usage{
setup_animate(path)
}
\arguments{
\item{path}{The path to the 'www' folder}
}
\description{
Copy the necessary JavaScript assets to the 'www' folder of a Shiny app
}
