% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{animate}
\alias{animate}
\title{A web-based graphics device}
\description{
A web-based graphics device

A web-based graphics device
}
\note{
This function differs from the \code{event} function in that events
registered through \code{simple_event} do not require R at deployment to work.
}
\examples{

## ------------------------------------------------
## Method `animate$chain`
## ------------------------------------------------

\dontrun{
library(animate)
device <- animate$new(600, 600)
attach(device)
par(xlim = c(0, 10), ylim = c(0, 10))
plot(1:10, 1:10, id = 1:10)
points(1:10, sample(10, 10), id = 1:10,
  transition = list(
    duration = 1000,
    on = chain(function(message) {
      print(message)
      points(1:10, sample(10, 10), id = 1:10, bg = "green",
             transition = list(duration = 2000))
      })
  ))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A handle for the WebSocket connection.}

\item{\code{ready_state}}{The ready state of the connection.}

\item{\code{shiny}}{Whether the device is used with shiny.}

\item{\code{session}}{A shiny session.}

\item{\code{virtual_meta}}{A list of device metadata.}

\item{\code{virtual_session}}{A virtual session simulated with V8.}

\item{\code{event_handlers}}{A named list of user-defined functions for handling events.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{animate$new()}}
\item \href{#method-off}{\code{animate$off()}}
\item \href{#method-send}{\code{animate$send()}}
\item \href{#method-set_max_stacksize}{\code{animate$set_max_stacksize()}}
\item \href{#method-svg}{\code{animate$svg()}}
\item \href{#method-bars}{\code{animate$bars()}}
\item \href{#method-objects}{\code{animate$objects()}}
\item \href{#method-plot}{\code{animate$plot()}}
\item \href{#method-points}{\code{animate$points()}}
\item \href{#method-lines}{\code{animate$lines()}}
\item \href{#method-abline}{\code{animate$abline()}}
\item \href{#method-axis}{\code{animate$axis()}}
\item \href{#method-text}{\code{animate$text()}}
\item \href{#method-image}{\code{animate$image()}}
\item \href{#method-event}{\code{animate$event()}}
\item \href{#method-chain}{\code{animate$chain()}}
\item \href{#method-simple_event}{\code{animate$simple_event()}}
\item \href{#method-set}{\code{animate$set()}}
\item \href{#method-par}{\code{animate$par()}}
\item \href{#method-remove}{\code{animate$remove()}}
\item \href{#method-clear}{\code{animate$clear()}}
\item \href{#method-delete}{\code{animate$delete()}}
\item \href{#method-import}{\code{animate$import()}}
\item \href{#method-export}{\code{animate$export()}}
\item \href{#method-record}{\code{animate$record()}}
\item \href{#method-observeAnimateEvent}{\code{animate$observeAnimateEvent()}}
\item \href{#method-clone}{\code{animate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$new(width, height, id = "SVG_1", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{An integer; the width in pixels.}

\item{\code{height}}{An integer; the height in pixels.}

\item{\code{id}}{A character string; the id assigned to the device.}

\item{\code{...}}{Additional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-off"></a>}}
\if{latex}{\out{\hypertarget{method-off}{}}}
\subsection{Method \code{off()}}{
Switch off the device; this function closes the WebSocket connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$off()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send commands to device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$send(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to send to the device.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_max_stacksize"></a>}}
\if{latex}{\out{\hypertarget{method-set_max_stacksize}{}}}
\subsection{Method \code{set_max_stacksize()}}{
Set the maximum size of the stack
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$set_max_stacksize(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The number of commands the plot stack can hold. Use
-1 for unlimited number of commands.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-svg"></a>}}
\if{latex}{\out{\hypertarget{method-svg}{}}}
\subsection{Method \code{svg()}}{
Initialise a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$svg(width = 800, height = 600, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{Width of the canvas in pixels.}

\item{\code{height}}{Height of the canvas in pixels.}

\item{\code{...}}{Additional parameters. Some commonly used parameters are \code{id}
and \code{root}. \code{id} assigns an id to the SVG element for future reference;
\code{root} specifies the DOM element to insert the SVG element into.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bars"></a>}}
\if{latex}{\out{\hypertarget{method-bars}{}}}
\subsection{Method \code{bars()}}{
Add bars to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$bars(x, y, w, h, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the bars.}

\item{\code{y}}{The y coordinates of the bars.}

\item{\code{w}}{The width of the bars.}

\item{\code{h}}{The height of the bars.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-objects"></a>}}
\if{latex}{\out{\hypertarget{method-objects}{}}}
\subsection{Method \code{objects()}}{
Add HTML objects to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$objects(x, y, w, h, content, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the objects.}

\item{\code{y}}{The y coordinates of the objects.}

\item{\code{w}}{The width of the objects.}

\item{\code{h}}{The height of the objects.}

\item{\code{content}}{The content of the objects; the HTML string.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Generic X-Y plotting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$plot(x, y, type = "p", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the data.}

\item{\code{y}}{The y coordinates of the data.}

\item{\code{type}}{Type of the plot; one of 'p' and 'l'.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points"></a>}}
\if{latex}{\out{\hypertarget{method-points}{}}}
\subsection{Method \code{points()}}{
Add points to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$points(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the points.}

\item{\code{y}}{The y coordinates of the points.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Options for the "pch" parameter: "circle", "plus", "diamond", "square",
"star", "triangle", "wye", "triangle_down", "triangle_left", "triangle_right",
"diamond_alt", "diamond_square", "pentagon", "hexagon", "hexagon_alt",
"octagon", "octagon_alt", "cross".

The unit of the "cex" parameter is squared pixels, corresponding to how
much pixel space the symbol would cover. The convention comes from the
d3 library, and the choice is (believed) to make plots visually consistent
across the different symbols.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lines"></a>}}
\if{latex}{\out{\hypertarget{method-lines}{}}}
\subsection{Method \code{lines()}}{
Add line segments / paths to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$lines(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the line.}

\item{\code{y}}{The y coordinates of the line.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-abline"></a>}}
\if{latex}{\out{\hypertarget{method-abline}{}}}
\subsection{Method \code{abline()}}{
Add straight lines to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$abline(a, b, h, v, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{The intercept.}

\item{\code{b}}{The slope.}

\item{\code{h}}{The y-value(s) for horizontal line(s).}

\item{\code{v}}{The x-value(s) for vertical line(s).}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-axis"></a>}}
\if{latex}{\out{\hypertarget{method-axis}{}}}
\subsection{Method \code{axis()}}{
Add an axis to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$axis(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the text.}

\item{\code{...}}{Additional graphical parameters.}

\item{\code{y}}{The y coordinates of the text.}

\item{\code{labels}}{The text.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\if{latex}{\out{\hypertarget{method-text}{}}}
\subsection{Method \code{text()}}{
Add text to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$text(x, y, labels, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the text.}

\item{\code{y}}{The y coordinates of the text.}

\item{\code{labels}}{The text.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image"></a>}}
\if{latex}{\out{\hypertarget{method-image}{}}}
\subsection{Method \code{image()}}{
Add background image to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$image(href, width, height, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{href}}{The link to the image.}

\item{\code{width}}{The width of the image.}

\item{\code{height}}{Th height of the image.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-event"></a>}}
\if{latex}{\out{\hypertarget{method-event}{}}}
\subsection{Method \code{event()}}{
Attach an interactive event to an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$event(selector, event_type, callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selector}}{A character string; a CSS selector.}

\item{\code{event_type}}{A character string; the event type. For example, "click", "mouseover",
"mouseout". See more options at \url{https://www.w3schools.com/jsref/dom_obj_event.asp}.}

\item{\code{callback}}{A function, to be called when the event is triggered.
The function should take an argument to receive the data from the
browser end.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chain"></a>}}
\if{latex}{\out{\hypertarget{method-chain}{}}}
\subsection{Method \code{chain()}}{
Chain a transition after another.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$chain(callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{callback}}{A function, to be called when the event is triggered.
The function should take an argument to receive the data from the
browser end.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
library(animate)
device <- animate$new(600, 600)
attach(device)
par(xlim = c(0, 10), ylim = c(0, 10))
plot(1:10, 1:10, id = 1:10)
points(1:10, sample(10, 10), id = 1:10,
  transition = list(
    duration = 1000,
    on = chain(function(message) {
      print(message)
      points(1:10, sample(10, 10), id = 1:10, bg = "green",
             transition = list(duration = 2000))
      })
  ))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simple_event"></a>}}
\if{latex}{\out{\hypertarget{method-simple_event}{}}}
\subsection{Method \code{simple_event()}}{
Attach a captured event to an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$simple_event(selector, event_type, method, param)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selector}}{A character string; a CSS selector.}

\item{\code{event_type}}{A character string; the event type. For example, "click", "mouseover",
"mouseout". See more options at \url{https://www.w3schools.com/jsref/dom_obj_event.asp}.}

\item{\code{method}}{A character string; the name of a device function (e.g. "points").}

\item{\code{param}}{A named list of arguments to be called with.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the active device to a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$set(device_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{device_id}}{A character vector; ID of the device.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par"></a>}}
\if{latex}{\out{\hypertarget{method-par}{}}}
\subsection{Method \code{par()}}{
Set the graphical parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$par(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The graphical parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
Remove elements from the active SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$remove(id = NULL, selector = "*")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A character vector; the ID of the elements.}

\item{\code{selector}}{A character vector; a CSS selector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all elements from the active SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Remove a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$delete(id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{A character string; the ID of the SVG. If not provided, remove
the active SVG element.
#' @description
#' Perform a group of graphical operations to a plot
#' @param ... Any number of graphical operations.
group = function(...) {
self$send(Message("fn_group", c(...)))
},}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$import(setting)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{setting}}{A JSON file exported from previous runs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$export(path = "./animate.json", handler = "browser")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A character string; the file path to export to.}

\item{\code{handler}}{'r' or 'browser'; the program to handle the export operation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-record"></a>}}
\if{latex}{\out{\hypertarget{method-record}{}}}
\subsection{Method \code{record()}}{
Record an animated plot as a MP4 video
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$record()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This function will prompt you to select a screen / window / tab to
record. Once started, the recording can be stopped by using the stop button
at the notification box, or clicking anywhere on the page near the device.
Always confirm that the screen recording notification box is gone.
The captured video will be downloaded right after the recording stops.

This uses web browsers' Media Streams API to record the screen
and return the captured frames as a video. The entire process runs locally.
The source file that provides this functionality can be found at
\code{system.file("addons/screen_record.js", package = "animate")}.

This function is disabled for Shiny and R Markdown Document.

This function does not work in the RStudio viewer. Please use the
"show in new window" button to launch the page with a web browser.

See browser compatibility at: \url{https://developer.mozilla.org/en-US/docs/Web/API/MediaStream_Recording_API#browser_compatibility}

See Media Streams API reference at: \url{https://developer.mozilla.org/en-US/docs/Web/API/Media_Streams_API}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-observeAnimateEvent"></a>}}
\if{latex}{\out{\hypertarget{method-observeAnimateEvent}{}}}
\subsection{Method \code{observeAnimateEvent()}}{
Event handler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$observeAnimateEvent(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{The input object in the \code{server} function of a Shiny app.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{animate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
