% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{click_to_loop}
\alias{click_to_loop}
\title{Click an element to play all frames}
\usage{
click_to_loop(selector = "#SVG_1", start = 2, wait = 20)
}
\arguments{
\item{selector}{The ID of the DOM element.}

\item{start}{An integer; the number of frames to execute upon the beginning
of the visualisation. This is useful when one wants to start with some set-up
instead of an empty canvas.}

\item{wait}{A number; the number of milliseconds to wait for before the
next frame is drawn.}
}
\description{
Click an element to play all frames
}
