% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{insert_animate}
\alias{insert_animate}
\title{Insert an animated plot into an R Markdown document}
\usage{
insert_animate(file)
}
\arguments{
\item{file}{The exported plot.}
}
\description{
Insert an animated plot into an R Markdown document
}
