% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{load_animate}
\alias{load_animate}
\title{Load the assets for an animate plot}
\usage{
load_animate(n = 0)
}
\arguments{
\item{n}{The number of commands the plot stack can hold. Use -1 for unlimited
number of commands.}
}
\description{
Load the assets for an animate plot
}
\note{
This function will only work if the 'www' folder of the Shiny app contains
the relevant JavaScript files. The files can be created using the function
\code{setup_animate}.
}
