% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{loop}
\alias{loop}
\title{Loop through the available frames n times}
\usage{
loop(times = 1, wait = 20)
}
\arguments{
\item{times}{An integer; the number of times to loop.}

\item{wait}{A number; the number of milliseconds to wait for before the
next frame is drawn.}
}
\description{
Loop through the available frames n times
}
