% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{plot2}
\alias{plot2}
\title{A web-based graphics device}
\description{
A web-based graphics device

A web-based graphics device
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection}}{A handle for the WebSocket connection.}

\item{\code{plot_stack}}{A stack to store the processed commands. This can be
exported to generate a standalone plot.}

\item{\code{stack_size}}{The number of commands \code{plot_stack} can hold. Use
-1 for unlimited number of commands.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{plot2$new()}}
\item \href{#method-off}{\code{plot2$off()}}
\item \href{#method-send}{\code{plot2$send()}}
\item \href{#method-init}{\code{plot2$init()}}
\item \href{#method-points}{\code{plot2$points()}}
\item \href{#method-lines}{\code{plot2$lines()}}
\item \href{#method-text}{\code{plot2$text()}}
\item \href{#method-import}{\code{plot2$import()}}
\item \href{#method-export}{\code{plot2$export()}}
\item \href{#method-clone}{\code{plot2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor of the device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$new(use_websocket = TRUE, import, stack_size = -1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_websocket}}{TRUE / FALSE; whether to enable websocket connection.}

\item{\code{import}}{A file path; needed when one wants to detach a plot from R.}

\item{\code{stack_size}}{An integer; the number of commands \code{plot_stack} can hold.
Use -1 for unlimited number of commands.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-off"></a>}}
\if{latex}{\out{\hypertarget{method-off}{}}}
\subsection{Method \code{off()}}{
Switch off the device; this function closes the WebSocket connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$off()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\if{latex}{\out{\hypertarget{method-send}{}}}
\subsection{Method \code{send()}}{
Send commands to device
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$send(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to send to the device.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
Initialise a SVG element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$init(cw = 800, ch = 600, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cw}}{Width of the canvas in pixels.}

\item{\code{ch}}{Height of the canvas in pixels.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-points"></a>}}
\if{latex}{\out{\hypertarget{method-points}{}}}
\subsection{Method \code{points()}}{
Add points to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$points(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the plot.}

\item{\code{y}}{The y coordinates of the plot.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lines"></a>}}
\if{latex}{\out{\hypertarget{method-lines}{}}}
\subsection{Method \code{lines()}}{
Add line segments / paths to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$lines(x, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the plot.}

\item{\code{y}}{The y coordinates of the plot.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-text"></a>}}
\if{latex}{\out{\hypertarget{method-text}{}}}
\subsection{Method \code{text()}}{
Add text to a plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$text(x, y, labels, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The x coordinates of the plot.}

\item{\code{y}}{The y coordinates of the plot.}

\item{\code{labels}}{The text.}

\item{\code{...}}{Additional graphical parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-import"></a>}}
\if{latex}{\out{\hypertarget{method-import}{}}}
\subsection{Method \code{import()}}{
Import an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$import()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\if{latex}{\out{\hypertarget{method-export}{}}}
\subsection{Method \code{export()}}{
Export an animated plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{plot2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
